POOja_README.md
I am Pooja rajpurohit, a highly dedicated and passionate individual with a strong connection to the computer and coding.A thrive on exploring new technologies and constantly seek to expand my new knowledge and skill in the field. I have a natural curiosity and love for the learning, always eager to take on new challenges and push myself to grow.